+++ HISTORY.md
@@ -6,6 +6,7 @@
 * In rare circumstances, `Simulation.run` targets can go out of date. Targets are now stored by-name.
 * Added `Sel.findFirstIn`.
 * `ReactTestUtils` now accept plain old `ReactElement`s.
+* Styles now given to React in camel case. No more warnings.
 
 # 0.6.0 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.4...v0.6.0))